+++ src/components/QueryFileUploader/utils/index.js
@@ -1,1 +1,2 @@
 export * from './generateFileDefinitionBody';
+export * from './checkFileHaveCsvExtension';